requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
requests==2.31.0
cryptography==41.0.7
google-generativeai==0.3.2
gunicorn==21.2.0
cryptography
Flask
google-generativeai
requests
Werkzeug
Pillow
Pillow